add bplist module
BP is short for BackPort. When you need to do a lot of backporting it
is useful to be able to mark/unmark and load/save lists of commits.

Currently tig has one global bplist, but the implementation is generic
enough and works on bplist instances so that in the future we could
have multiple bplist.

* bplist_init(): initialize a bplist
* bplist_read(): load a bplist from a file into an initialized bplist
* bplist_has_rev(): checks whether a bplist contains a commit rev
* bplist_add_rev(): adds a rev to a bplist if it is not already there
* bplist_rem_rev(): removes a rev from a bplist if it holds it
* bplist_toggle_rev(): adds/remove rev from a bplist
* bplist_write(): dump a bplist to a file

A bplist file is a plain text file where each line is in the form of

    <sha1>[ <text>]

Lines who do not match this format will be added as-is and not be
considered as commit. When writing a bplist, commits are sorted by
commit date.

Non-commit lines get a commit date of 0 and they end up at the top of
the line as a result.